Updated documentation and some path references
Documentation has been updated to reflect new data references.
Changed some path references in diagnostics so that they are
relative to the directory containing the respective diagnostic
script.
Added a copy of the xcode project file/folder for use by others.
This avoids issues related to changes in this file every time it is
opened.
Updated the main readme to describe this new version.
Updated example_outputs for aez boundary example.

 Changes to be committed:
	modified:   README.md
	modified:   diagnostics/Carbon_diagnostic_functions.R
	modified:   diagnostics/Compare_LDS_area_outputs.R
	modified:   diagnostics/generate_moirai_spatial.R
	modified:   diagnostics/plot_moirai_crop_ctry.r
	modified:   diagnostics/plot_moirai_crop_rglu.r
	modified:   diagnostics/plot_moirai_land_type_area.r
	modified:   diagnostics/plot_moirai_landrent_glu.r
	modified:   diagnostics/readme.md
	modified:   docs/moirai_v31_table_1.pdf
	modified:   docs/third_party_contributions_v31.pdf
	modified:   example_outputs/aez_orig/Ref_veg_carbon_Mg_per_ha.csv
	modified:   example_outputs/aez_orig/aglu-data/moirai/Ref_veg_carbon_Mg_per_ha.csv
	modified:   example_outputs/aez_orig/moirai_log_aez_orig.txt
	modified:   input_files/moirai_input_aez_orig.txt
	new file:   moirai.xcodeproj.copy/